requirements.txt
langchain>=1.0.0
langchain-core>=1.0.0
langchain-community>=0.4.0
langchain-groq>=1.0.0
langchain-chroma>=1.0.0
chromadb>=0.4.24
fastembed>=0.2.6
streamlit>=1.32.0
beautifulsoup4>=4.12.3
python-dotenv>=1.0.1
pandas>=2.2.1
